% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_nops_eval.R
\name{browse_nops_eval}
\alias{browse_nops_eval}
\title{Launches a shiny gadget for browsing nops_eval.zip of written
exams created with R/exams.}
\usage{
browse_nops_eval(file, width = 1000, height = 800)
}
\arguments{
\item{file}{character. Path to nops_eval.zip or unzipped directory.}

\item{width}{integer. Width of dialogViewer}

\item{height}{integer. Height of dialogViewer}
}
\value{
data.frame that will be used to overwrite Data.txt
}
\description{
Launches a shiny gadget for browsing nops_eval.zip of written
exams created with R/exams.
}
