% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course.R
\name{course_init}
\alias{course_init}
\title{Initialize course directory.}
\usage{
course_init()
}
\value{
Write out YAML file (course.yml) that collects course information.
}
\description{
Initialize course directory.
}
\examples{
\dontrun{
  course_init()
}
}
