% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam}
\alias{exam}
\title{Set up one exam.}
\usage{
exam(problems, date, name, shortname = date, versions = 1,
  samples.from.group = 1)
}
\arguments{
\item{problems}{list of problems.}

\item{date}{character. Representing the date of exam. e.g.) "2019-10-10"}

\item{name}{character. Name of the exam. e.g.) "Final Exam"}

\item{shortname}{character. Short name for the exam. e.g.) "Day02" or "final"}

\item{versions}{integer. Number of versions created.}

\item{samples.from.group}{integer vector. a.k.a \code{nsamp} in R/exams world.}
}
\value{
list of exam information. Result is passed to \code{exam_create}
}
\description{
Set up one exam.
}
