% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam_create}
\alias{exam_create}
\title{Step 1: Create NOPS written exam}
\usage{
exam_create(exam, latex.cover = NULL, html.template = NULL,
  encryptHTML = FALSE, course.as.pdf.title = TRUE, ...)
}
\arguments{
\item{exam}{list of exam information. Created with \code{exam}}

\item{latex.cover}{character. Path to a tex file cover page or intro message.
Passed to \code{intro} parameter of \code{exams2nops}}

\item{html.template}{character. Path to template html. Passed to \code{template}
parameter of \code{exams2html}}

\item{encryptHTML}{logical or character. If TRUE or passphrase is set, password-protected
version of HTML problem set is also created.}

\item{course.as.pdf.title}{logical. If TRUE, use course title as PDF title. If FALSE,
institution is used.}

\item{..., }{other parameters passed to \code{exams2nops}}
}
\value{
invisible NULL
}
\description{
Step 1: Create NOPS written exam
}
