% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam_evaluate}
\alias{exam_evaluate}
\title{Step 5: Evaluate the exam}
\usage{
exam_evaluate(
  shortname,
  flavor = "template",
  report.template = system.file("nops", "eval.html", package = "exams"),
  eval = exams::exams_eval(partial = FALSE, negative = FALSE),
  ...
)
}
\arguments{
\item{shortname}{short name, directory contains config.rds, created with \code{exam}}

\item{flavor}{character. User nops_eval_write_<flavor>}

\item{report.template}{character. Passed on to nops_eval_write_<flavor>
as an argument to template paramter}

\item{eval}{list specification of evaluation policy}

\item{...}{Parameters passed on to \code{nops_eval} and \code{nops_eval_write_*}.
As for \code{nops_eval}, respected parameters are \code{points}, \code{mark},
\code{labels}, \code{results}, \code{file}, \code{interactive},
\code{string_scans}, \code{string_points}. The other parameters are ignored.}
}
\value{
\code{data.frame} of marked result.
}
\description{
Step 5: Evaluate the exam
}
