% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam_scan}
\alias{exam_scan}
\title{Step 2: Print
Step 3: Exam
Step 4: Scan exam sheets}
\usage{
exam_scan(exam, images, ...)
}
\arguments{
\item{exam}{character. path to saved RDS file.}

\item{images}{scanned images (PNG's or PDF file)}

\item{...}{parameters passed on to \code{exams::nops_scan}}
}
\value{
contents of Daten.txt as a character vector. Returned invisibly.
}
\description{
Step 2: Print
Step 3: Exam
Step 4: Scan exam sheets
}
