% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nops_eval_browse.R
\name{nops_eval_browse}
\alias{nops_eval_browse}
\title{Launches a shiny gadget for browsing nops_eval.zip of written
exams created with R/exams.}
\usage{
nops_eval_browse(file, width = 1000, height = 800)
}
\arguments{
\item{file}{Path to nops_eval.zip or unzipped directory.}

\item{width}{Width of dialogViewer}

\item{height}{Height of dialogViewer}
}
\description{
Launches a shiny gadget for browsing nops_eval.zip of written
exams created with R/exams.
}
