% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_nops_results.R
\name{nops_eval_listview}
\alias{nops_eval_listview}
\title{Serve listview of nops_eval.zip}
\usage{
nops_eval_listview(name)
}
\arguments{
\item{name}{Path to output zip file of nops_eval()}
}
\value{
Directory name to serve, invisibly
}
\description{
Serve listview of nops_eval.zip
}
