% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nops_eval_write.R
\name{nops_eval_write_examtools}
\alias{nops_eval_write_examtools}
\title{Customized nops_eval_write for examtools}
\usage{
nops_eval_write_examtools(
  results = "nops_eval.csv",
  name = "report.html",
  template = NULL,
  ...
)
}
\arguments{
\item{results}{character or data.frame. Path to nops_eval.csv.}

\item{name}{character. Output file name for individual reports, with extension.}

\item{template}{character. HTML template for output files.}

\item{...}{list. \code{encoding}, \code{language}, \code{converter}, and \code{dir}, passed to
\code{\link[exams]{nops_eval}}, are respected.}
}
\description{
Customized nops_eval_write for examtools
}
