% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problems.R
\name{problems_list}
\alias{problems_list}
\title{Lists problems in a directory.}
\usage{
problems_list(path, group.sep = "-", exclude = NULL, basedir = NULL)
}
\arguments{
\item{path}{character. Path to a directory that contains problems Rnw/Rmd, relative to basedir.
This function assumes particular naming rules. For further details, see below.}

\item{group.sep}{character. Character to indicate groups. For further details, see below.}

\item{exclude}{character string containing a regular expression to be searched for.
Files that match this pattern are excluded from the output.}

\item{basedir}{character. Base directory relative to which Path is specified. Defaults to
problems_dir in course.yml}
}
\value{
List, which can be passed to exams function such as \code{exams::exams2pdf}.
}
\description{
Lists problems in a directory.
}
\details{
Problem filen ames should be like \code{01-abc.Rmd}, \code{02-xyz.Rmd},
  where \code{-} is the character separating group (\code{01}, \code{02}) and
  keyword (\code{abc} and \code{xyz}). You can use different \code{group.sep}
  than \code{-} (default) by specifying \code{group.sep} argument.

  When the directory specified by the \code{path} argument contains
  \code{01-abc.Rmd}, \code{01-def.Rmd} and \code{02-xyz.Rmd}, this function
  returns \code{list(c("01-abc.Rmd", "01-def.Rmd"), "02-xyz.Rmd")}.
}
\examples{
\dontrun{
  path <- system.file("inst", "samples", package = "examtools")
  problems_list(path)
}
}
