% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.R
\name{rewrite_for_moodle}
\alias{rewrite_for_moodle}
\title{Convert structure of nops_eval.zip for Moodle
worksheet will also be updated.}
\usage{
rewrite_for_moodle(nops_zip, worksheet_csv, nops_csv = NULL,
  reglength = 7, suffix = "-moodle", worksheet_points_col = "Grade",
  pattern = "\%s_\%s_assignsubmission_file_", quote = TRUE)
}
\arguments{
\item{quote}{}
}
\description{
Convert structure of nops_eval.zip for Moodle
worksheet will also be updated.
}
