% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_daten.R
\name{check_daten}
\alias{check_daten}
\title{A shiny gadget to check and fix Daten.txt}
\usage{
check_daten(
  dir,
  file = "Daten.txt",
  width = 800L,
  height = 600L,
  n = 45L,
  downsample = 4L,
  nalt = 5L
)
}
\arguments{
\item{dir}{character. Path to nops_scan (uncompressed directory)}

\item{file}{character. Typically "Daten.txt"}

\item{width}{integer. Width of the dialog viewer.}

\item{height}{integer. Width of the dialog viewer.}

\item{n}{integer. Number of problems.}

\item{downsample}{integer. Image downsampling factor.}

\item{nalt}{integer. Maximum number of alternatives.}
}
\description{
A shiny gadget to check and fix Daten.txt
}
