% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{exam_scan}
\alias{exam_scan}
\title{Step 4: Scan exam sheets}
\usage{
exam_scan(exam, images, ...)
}
\arguments{
\item{exam}{short name, directory contains config.rds, created with \code{exam}}

\item{images}{scanned images (PNG's or PDF file)}

\item{...}{parameters passed on to \code{exams::nops_scan}. NB: \code{dir} and \code{file}
are overwritten by default configurations}
}
\value{
contents of Daten.txt as a character vector. Returned invisibly.
}
\description{
Step 4: Scan exam sheets
}
