% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nops_eval_write.R
\name{nops_eval_write_custom}
\alias{nops_eval_write_custom}
\title{Customized nops_eval_write}
\usage{
nops_eval_write_custom(results = "nops_eval.csv",
  file = "nops_eval.zip", name = "report.html", template = NULL,
  encoding = "UTF-8", language = "en", converter = NULL, ...)
}
\arguments{
\item{results}{character. Path to nops_eval.csv}

\item{file}{character. Path to nops_eval.zip}

\item{name}{character. Output file name, with extension.}

\item{template}{character. HTML template for output files.}

\item{encoding}{character. Encoding as in \code{\link[exams]{nops_eval}}}

\item{language}{character. Language as in \code{\link[exams]{nops_eval}}}

\item{converter}{character. Converter as in \code{\link[exams]{nops_eval}}}

\item{...}{list. Ignored.}
}
\description{
Customized nops_eval_write
}
